Add storage service and helper functions for toss management
