+++ examples/xmtp-group-toss/index.ts
@@ -1,6 +1,4 @@
-import { existsSync, mkdirSync } from "fs";
 import * as fs from "fs/promises";
-import * as path from "path";
 import {
   AgentKit,
   cdpApiActionProvider,
@@ -15,7 +13,6 @@ import {
   Wallet,
   type Transfer as CoinbaseTransfer,
   type Trade,
-  type WalletData,
 } from "@coinbase/coinbase-sdk";
 import { HumanMessage } from "@langchain/core/messages";
 import { MemorySaver } from "@langchain/langgraph";
@@ -31,13 +28,89 @@ import { isAddress } from "viem";
 import "dotenv/config";
 import { createSigner, getEncryptionKeyFromHex } from "@helpers/client";
 import { logAgentDetails, validateEnvironment } from "@helpers/utils";
+import {
+  extractJsonFromResponse,
+  initializeCoinbaseSDK,
+  storage,
+  TossStatus,
+  XMTP_STORAGE_DIR,
+  type AgentConfig,
+  type AgentWalletData,
+  type GroupTossName,
+  type ParsedToss,
+  type StreamChunk,
+  type TossJsonResponse,
+  type TransferResponse,
+} from "./helper";
 
-const { CDP_API_KEY_NAME, CDP_API_KEY_PRIVATE_KEY, NETWORK_ID } =
-  validateEnvironment([
-    "CDP_API_KEY_NAME",
-    "CDP_API_KEY_PRIVATE_KEY",
-    "NETWORK_ID",
-  ]);
+// Global stores for memory and agent instances
+const memoryStore: Record<string, MemorySaver> = {};
+const agentStore: Record<string, ReturnType<typeof createReactAgent>> = {};
+
+const {
+  CDP_API_KEY_NAME,
+  CDP_API_KEY_PRIVATE_KEY,
+  NETWORK_ID,
+  XMTP_ENV,
+  WALLET_KEY,
+  ENCRYPTION_KEY,
+} = validateEnvironment([
+  "CDP_API_KEY_NAME",
+  "CDP_API_KEY_PRIVATE_KEY",
+  "NETWORK_ID",
+  "XMTP_ENV",
+  "WALLET_KEY",
+  "ENCRYPTION_KEY",
+]);
+
+// Constants for default values
+const DEFAULT_OPTIONS = ["yes", "no"];
+const DEFAULT_AMOUNT = "1";
+const USDC_TOKEN_ADDRESS = "0x5dEaC602762362FE5f135FA5904351916053cF70";
+
+/**
+ * Agent instruction template for coin toss activities
+ */
+const AGENT_INSTRUCTIONS = `
+  You are a CoinToss Agent that helps users participate in coin toss activities.
+  
+  You have two main functions:
+  1. Process natural language toss requests and structure them
+  2. Handle coin toss management commands
+  
+  When parsing natural language tosses:
+  - Extract the toss topic (what people are tossing on)
+  - Identify options (default to "yes" and "no" if not provided)
+  - Determine toss amount (default to 1 USDC if not specified)
+  - Enforce a maximum toss amount of 10 USDC
+  
+  For example:
+  - "Will it rain tomorrow for 5" should be interpreted as a toss on "Will it rain tomorrow" with options ["yes", "no"] and amount "5"
+  - "Lakers vs Celtics for 10" should be interpreted as a toss on "Lakers vs Celtics game" with options ["Lakers", "Celtics"] and amount "10"
+  
+  When checking payments or balances:
+  1. Use the USDC token at ${USDC_TOKEN_ADDRESS} on Base.
+  2. When asked to check if a payment was sent, verify:
+     - The exact amount was transferred
+     - The transaction is confirmed
+     - The correct addresses were used
+  3. For balance checks, show the exact USDC amount available.
+  4. When transferring winnings, ensure:
+     - The toss wallet has sufficient balance
+     - The transfer is completed successfully
+     - Provide transaction details
+  
+  Available commands:
+  @toss <topic> <options> <amount> - Create a new toss
+  /join <tossId> <option> - Join an existing toss with the specified ID
+  /close <tossId> <option> - Close the toss and set the winning option (creator only)
+  /status <tossId> - Check toss status and participants
+  /list - List all active tosses
+  /balance - Check your wallet balance
+  /help - Show available commands
+  
+  Keep responses concise and clear, focusing on payment verification and toss status.
+`;
 
 // Constants
 export const HELP_MESSAGE = `Available commands:
@@ -53,22 +126,6 @@ Other commands:
 @toss close <tossId> <option> - Close the toss and set the winning option (only for toss creator)
 @toss help - Show this help message
 `;
-
-// Initialize Coinbase SDK
-function initializeCoinbaseSDK(): boolean {
-  try {
-    Coinbase.configure({
-      apiKeyName: CDP_API_KEY_NAME,
-      privateKey: CDP_API_KEY_PRIVATE_KEY,
-    });
-    console.log("Coinbase SDK initialized successfully, network:", NETWORK_ID);
-    return true;
-  } catch (error: unknown) {
-    const errorMessage = error instanceof Error ? error.message : String(error);
-    console.error("Failed to initialize Coinbase SDK:", errorMessage);
-    return false;
-  }
-}
 
 // Initialize the SDK when the module is loaded
 let sdkInitialized = false;
@@ -715,202 +772,6 @@ async function handleNaturalLanguageCommand(
   return response;
 }
 
-export const WALLET_STORAGE_DIR = ".data/wallet_data";
-export const XMTP_STORAGE_DIR = ".data/xmtp";
-export const TOSS_STORAGE_DIR = ".data/tosses";
-
-/**
- * Storage service for coin toss  data and user wallets
- */
-class StorageService {
-  private initialized = false;
-
-  constructor() {
-    // Initialize directories on creation
-    this.initialize();
-  }
-
-  /**
-   * Initialize storage directories
-   */
-  public initialize(): void {
-    if (this.initialized) return;
-
-    // Ensure storage directories exist
-    [WALLET_STORAGE_DIR, TOSS_STORAGE_DIR, XMTP_STORAGE_DIR].forEach((dir) => {
-      if (!existsSync(dir)) {
-        mkdirSync(dir, { recursive: true });
-      }
-    });
-
-    this.initialized = true;
-    console.log("Local file storage initialized");
-  }
-
-  /**
-   * Save data to a JSON file
-   */
-  private async saveToFile(
-    directory: string,
-    identifier: string,
-    data: string,
-  ): Promise<boolean> {
-    const toRead = `${identifier}-${NETWORK_ID}`;
-    try {
-      const filePath = path.join(directory, `${toRead}.json`);
-      await fs.writeFile(filePath, data);
-      return true;
-    } catch (error) {
-      console.error(`Error writing to file ${toRead}:`, error);
-      return false;
-    }
-  }
-
-  /**
-   * Read data from a JSON file
-   */
-  private async readFromFile<T>(
-    directory: string,
-    identifier: string,
-  ): Promise<T | null> {
-    try {
-      const key = `${identifier}-${NETWORK_ID}`;
-      const filePath = path.join(directory, `${key}.json`);
-      const data = await fs.readFile(filePath, "utf-8");
-      return JSON.parse(data) as T;
-    } catch (error) {
-      // If file doesn't exist, return null
-      if (
-        error instanceof Error &&
-        (error.message.includes("ENOENT") ||
-          error.message.includes("no such file or directory"))
-      ) {
-        return null;
-      }
-      // For other errors, rethrow
-      throw error;
-    }
-  }
-
-  /**
-   * Save a coin toss game
-   */
-  public async saveToss(toss: GroupTossName): Promise<void> {
-    if (!this.initialized) this.initialize();
-    await this.saveToFile(TOSS_STORAGE_DIR, toss.id, JSON.stringify(toss));
-  }
-
-  /**
-   * Get a coin toss game by ID
-   */
-  public async getToss(tossId: string): Promise<GroupTossName | null> {
-    if (!this.initialized) this.initialize();
-    return this.readFromFile<GroupTossName>(TOSS_STORAGE_DIR, tossId);
-  }
-
-  /**
-   * List all active games
-   */
-  public async listActiveTosses(): Promise<GroupTossName[]> {
-    if (!this.initialized) this.initialize();
-
-    const tosses: GroupTossName[] = [];
-    try {
-      const files = await fs.readdir(TOSS_STORAGE_DIR);
-
-      for (const file of files) {
-        if (file.endsWith(".json")) {
-          const tossId = file.replace(`-${NETWORK_ID}.json`, "");
-          const toss = await this.getToss(tossId);
-          if (
-            toss &&
-            toss.status !== TossStatus.COMPLETED &&
-            toss.status !== TossStatus.CANCELLED
-          ) {
-            tosses.push(toss);
-          }
-        }
-      }
-    } catch (error) {
-      console.error("Error listing active games:", error);
-    }
-
-    return tosses;
-  }
-
-  /**
-   * Update an existing game (alias for saveToss)
-   */
-  public async updateToss(toss: GroupTossName): Promise<void> {
-    await this.saveToss(toss);
-  }
-
-  /**
-   * Save user wallet data
-   */
-  public async saveWallet(inboxId: string, walletData: string): Promise<void> {
-    if (!this.initialized) this.initialize();
-    await this.saveToFile(WALLET_STORAGE_DIR, inboxId, walletData);
-  }
-
-  /**
-   * Get user wallet data by user ID
-   */
-  public async getWallet(inboxId: string): Promise<AgentWalletData | null> {
-    if (!this.initialized) this.initialize();
-    return this.readFromFile<AgentWalletData>(WALLET_STORAGE_DIR, inboxId);
-  }
-
-  /**
-   * Delete a file
-   */
-  public async deleteFile(directory: string, key: string): Promise<boolean> {
-    try {
-      const filePath = path.join(directory, `${key}.json`);
-      await fs.unlink(filePath);
-      return true;
-    } catch (error) {
-      console.error(`Error deleting file ${key}:`, error);
-      return false;
-    }
-  }
-
-  /**
-   * Get wallet count
-   */
-  public async getWalletCount(): Promise<number> {
-    try {
-      const files = await fs.readdir(WALLET_STORAGE_DIR);
-      return files.filter((file) => file.endsWith(".json")).length;
-    } catch (error) {
-      console.error("Error getting wallet count:", error);
-      return 0;
-    }
-  }
-
-  /**
-   * Get the toss storage directory
-   */
-  public getTossStorageDir(): string {
-    return TOSS_STORAGE_DIR;
-  }
-}
-
-// Create a single global instance
-const storage = new StorageService();
-
-// Export the storage instance
-export { storage };
-
-// Interface for transfer response
-interface Transfer {
-  model?: {
-    sponsored_send?: {
-      transaction_link?: string;
-    };
-  };
-}
-
 export class TossManager {
   private walletService: WalletService;
 
@@ -1218,7 +1079,7 @@ export class TossManager {
 
           // Set transaction link from first successful transfer
           if (!toss.transactionLink) {
-            const transferData = transfer as unknown as Transfer;
+            const transferData = transfer as unknown as TransferResponse;
             toss.transactionLink =
               transferData.model?.sponsored_send?.transaction_link;
           }
@@ -1308,94 +1169,6 @@ export class TossManager {
     return toss;
   }
 }
-
-// Interface to track participant options
-export interface Participant {
-  inboxId: string;
-  option: string;
-}
-
-export interface GroupTossName {
-  id: string;
-  creator: string;
-  tossAmount: string;
-  status: TossStatus;
-  participants: string[]; // Maintaining for backward compatibility
-  participantOptions: Participant[]; // New field to track participant options
-  winner?: string;
-  walletAddress: string;
-  createdAt: number;
-  tossResult?: string;
-  paymentSuccess?: boolean;
-  transactionLink?: string;
-  tossTopic?: string;
-  tossOptions?: string[];
-}
-
-export enum TossStatus {
-  CREATED = "CREATED",
-  WAITING_FOR_PLAYER = "WAITING_FOR_PLAYER",
-  READY = "READY",
-  IN_PROGRESS = "IN_PROGRESS",
-  COMPLETED = "COMPLETED",
-  CANCELLED = "CANCELLED",
-}
-
-// Agent wallet data
-export type AgentWalletData = {
-  id: string;
-  walletData: WalletData;
-  agent_address: string;
-  inboxId: string;
-  wallet?: Wallet;
-};
-
-export interface AgentConfig {
-  configurable: {
-    thread_id: string;
-  };
-}
-
-// Interface for parsed toss information
-export interface ParsedToss {
-  topic: string;
-  options: string[];
-  amount: string;
-}
-
-// Define stream chunk types
-export interface AgentChunk {
-  agent: {
-    messages: Array<{
-      content: string;
-    }>;
-  };
-}
-
-export interface ToolsChunk {
-  tools: {
-    messages: Array<{
-      content: string;
-    }>;
-  };
-}
-
-export type StreamChunk = AgentChunk | ToolsChunk;
-
-// Interface for parsed JSON response
-export interface TossJsonResponse {
-  topic?: string;
-  options?: string[];
-  amount?: string;
-  valid?: boolean;
-  reason?: string;
-}
-
-const { WALLET_KEY, ENCRYPTION_KEY, XMTP_ENV } = validateEnvironment([
-  "WALLET_KEY",
-  "ENCRYPTION_KEY",
-  "XMTP_ENV",
-]);
 
 /**
  * Initialize the XMTP client
@@ -1489,59 +1262,6 @@ export function extractCommand(content: string): string | null {
 
   return null;
 }
-
-// Constants for default values
-const DEFAULT_OPTIONS = ["yes", "no"];
-const DEFAULT_AMOUNT = "1";
-const USDC_TOKEN_ADDRESS = "0x5dEaC602762362FE5f135FA5904351916053cF70";
-
-/**
- * Agent instruction template for coin toss activities
- */
-const AGENT_INSTRUCTIONS = `
-  You are a CoinToss Agent that helps users participate in coin toss activities.
-  
-  You have two main functions:
-  1. Process natural language toss requests and structure them
-  2. Handle coin toss management commands
-  
-  When parsing natural language tosses:
-  - Extract the toss topic (what people are tossing on)
-  - Identify options (default to "yes" and "no" if not provided)
-  - Determine toss amount (default to 1 USDC if not specified)
-  - Enforce a maximum toss amount of 10 USDC
-  
-  For example:
-  - "Will it rain tomorrow for 5" should be interpreted as a toss on "Will it rain tomorrow" with options ["yes", "no"] and amount "5"
-  - "Lakers vs Celtics for 10" should be interpreted as a toss on "Lakers vs Celtics game" with options ["Lakers", "Celtics"] and amount "10"
-  
-  When checking payments or balances:
-  1. Use the USDC token at ${USDC_TOKEN_ADDRESS} on Base.
-  2. When asked to check if a payment was sent, verify:
-     - The exact amount was transferred
-     - The transaction is confirmed
-     - The correct addresses were used
-  3. For balance checks, show the exact USDC amount available.
-  4. When transferring winnings, ensure:
-     - The toss wallet has sufficient balance
-     - The transfer is completed successfully
-     - Provide transaction details
-  
-  Available commands:
-  @toss <topic> <options> <amount> - Create a new toss
-  /join <tossId> <option> - Join an existing toss with the specified ID
-  /close <tossId> <option> - Close the toss and set the winning option (creator only)
-  /status <tossId> - Check toss status and participants
-  /list - List all active tosses
-  /balance - Check your wallet balance
-  /help - Show available commands
-  
-  Keep responses concise and clear, focusing on payment verification and toss status.
-`;
-
-// Global stores for memory and agent instances
-const memoryStore: Record<string, MemorySaver> = {};
-const agentStore: Record<string, ReturnType<typeof createReactAgent>> = {};
 
 export async function initializeAgent(inboxId: string) {
   try {
@@ -1641,24 +1361,6 @@ export async function processMessage(
   } catch (error) {
     console.error("Error processing message:", error);
     return "Sorry, I encountered an error while processing your request. Please try again.";
-  }
-}
-
-/**
- * Extract JSON from agent response text
- * @param response The text response from agent
- * @returns Parsed JSON object or null if not found
- */
-function extractJsonFromResponse(response: string): TossJsonResponse | null {
-  try {
-    const jsonMatch = response.match(/\{[\s\S]*\}/);
-    if (jsonMatch) {
-      return JSON.parse(jsonMatch[0]) as TossJsonResponse;
-    }
-    return null;
-  } catch (error) {
-    console.error("Error parsing JSON from agent response:", error);
-    return null;
   }
 }
